+++ autodocodec/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.3.0.0] - 2024-06-16
+
+### Changed
+
+* Refactored `Codec` so it's input and output parameters have a representative, not nominal role.
+  This means one can now use `deriving newtype` with the `HasCodec` class.
+
 ## [0.2.3.0] - 2024-06-23
 
 ### Added